Added description to code and simplified the file name
